requirements.txt
pillow==9
brother_ql
python-dotenv
unificontrol
curlify